+++ src/core/mod.rs
@@ -2,4 +2,5 @@ pub mod checks;
 pub mod config;
 pub mod dbt_objects;
 pub mod manifest;
+pub mod run;
 pub mod traits;
